Get names for uuids.

Get names for uuids.

get_names(
  uuid,
  supertaxa = NULL,
  subtaxa = NULL,
  options = NULL,
  stripauthority = TRUE,
  ...
)

Arguments

uuid

UUID to get names for

supertaxa

If immediate, returns data for immediate supertaxa ("parents"). If all, returns data for all supertaxa ("ancestors"). Otherwise, does not include supertaxa.

subtaxa

If immediate, returns data for immediate subtaxa ("children"). Otherwise, does not include subtaxa.

options

See details for the options for options, get it, ha.

stripauthority

If TRUE (default) the authority is stripped off of the scientific name.

...

curl options passed on to crul::HttpClient

Details

Here are the options for the options argument:

citationStart: (optional) Integer Indicates where in the string the citation starts. May be null.

html: (optional) StringHTML version of the name.

namebankID: (optional) StringuBio Namebank identifier. May be null.

root: (optional) Boolean If true, this name has no hyperonyms (names of supertaxa). (Should only be true for Panbiota/Vitae.)

string: (optional) String The text of the name, including the citation, if any.

type: (optional) String Either "scientific or "vernacular.

uid: (always) String Universally unique identifier.

uri: (optional) String The unique URI associated with the name.

votes: (optional) Integer The number of votes this name has received. (Currently unused.)

Examples

if (FALSE) {
get_names(uuid = "f3254fbd-284f-46c1-ae0f-685549a6a373", options = "string")
get_names(uuid = "f3254fbd-284f-46c1-ae0f-685549a6a373", supertaxa="immediate", 
   options=c("string namebankID"))
get_names(uuid = "f3254fbd-284f-46c1-ae0f-685549a6a373", supertaxa="all", 
   options="string")
}